==> etapa5/src/2.FE_v1.ipynb <==
# Solución etapa 4 - FE + Training

# Utilidades para print
from utils.print_utils import tabl, headr, titl

# Cargar los datasets
import pandas as pd

historicos_ordenes_cleaned = pd.read_csv('../data/cleaned/historicos_ordenes_cleaned.csv')
registros_condiciones_cleaned = pd.read_csv('../data/cleaned/registros_condiciones_cleaned.csv')
caracteristicas_equipos_cleaned = pd.read_csv('../data/cleaned/caracteristicas_equipos_cleaned.csv')

## Feature eng

### Target

# Generar target

target_column="Fallo"

def code_target(val):
    if val == 'Correctivo':
        return 1
    else:
        return 0


historicos_ordenes_cleaned[target_column] = historicos_ordenes_cleaned["Tipo_Mantenimiento"].apply(code_target)
historicos_ordenes_cleaned.drop("Tipo_Mantenimiento", axis=1, inplace=True)

### historicos_ordenes

tabl(historicos_ordenes_cleaned)

historicos_ordenes_cleaned['Fecha'] = pd.to_datetime(historicos_ordenes_cleaned['Fecha'])

# historicos_ordenes_cleaned['Año'] = historicos_ordenes_cleaned['Fecha'].dt.year
# historicos_ordenes_cleaned['Mes'] = historicos_ordenes_cleaned['Fecha'].dt.month
# historicos_ordenes_cleaned['Día'] = historicos_ordenes_cleaned['Fecha'].dt.day

tabl(historicos_ordenes_cleaned)

### registros_condiciones

registros_condiciones_cleaned.drop(columns=['ID_Registro'], index=1, inplace=True)

registros_condiciones_cleaned['Fecha']=pd.to_datetime(registros_condiciones_cleaned['Fecha'])
# registros_condiciones_cleaned['Año'] = registros_condiciones_cleaned['Fecha'].dt.year
# registros_condiciones_cleaned['Mes'] = registros_condiciones_cleaned['Fecha'].dt.month
# registros_condiciones_cleaned['Día'] = registros_condiciones_cleaned['Fecha'].dt.day

tabl(registros_condiciones_cleaned)

### Unir registro + hitorico + caracteristicas

# Unir datasets
final_data = pd.merge(registros_condiciones_cleaned, historicos_ordenes_cleaned, on=['ID_Equipo','Fecha'], how='left')
final_data = pd.merge(final_data, caracteristicas_equipos_cleaned, on='ID_Equipo', how='inner')

tabl(final_data)

final_data.columns

# Limpiar columnas accesorias
final_data.drop(columns=['ID_Orden','Costo_Mantenimiento', 'Duracion_Horas', 'Ubicacion'], inplace=True)

tabl(final_data)

### Duplicados

# Eliminar duplicados

print(headr('Duplicados final_data:'),final_data.duplicated().sum())
# final_data.drop_duplicates(inplace=True)
# tabl(final_data)

# Verificamos nulos
print(headr('Nulos final_data:'),final_data.isna().sum())

columnas_con_nulos = final_data.columns[final_data.isna().any()].tolist()
columnas_con_nulos

# Imputamos nulos
from sklearn.impute import SimpleImputer

imputer = SimpleImputer(strategy="constant", fill_value=0)

final_data[columnas_con_nulos] = imputer.fit_transform(final_data[columnas_con_nulos])

tabl(final_data)

print("% fallos:",
final_data[final_data['Fallo'] == 1].shape[0] / final_data.shape[0] * 100)

final_data.info()

### Volvemos a explorar con columnas codificadas


# Extraer columnas numéricas y categóricas
from sklearn.compose import make_column_selector as selector

numerical_columns_selector = selector(dtype_exclude=object)
categorical_columns_selector = selector(dtype_include=object)

numerical_columns = numerical_columns_selector(final_data)
categorical_columns = categorical_columns_selector(final_data)

print(headr("Numerical columns"), numerical_columns)
print(headr("Categorical columns"), categorical_columns)

from sklearn.preprocessing import LabelEncoder, StandardScaler

final_data_coded = final_data.copy()
final_data_coded[categorical_columns] = final_data_coded[categorical_columns].apply(LabelEncoder().fit_transform)

# Eliminamos fecha para la evaluación
final_data_coded.drop(columns=['Fecha'], inplace=True)

tabl(final_data_coded)

# Evaluamos la aportación de cada columna
from sklearn.feature_selection import SelectKBest
from sklearn.feature_selection import f_classif,chi2

X= final_data_coded.drop(target_column, axis=1)
y = final_data_coded[target_column]

fvalue_selector = SelectKBest(f_classif, k=2)

X_kbest = fvalue_selector.fit(X,y)

feature_scores = pd.DataFrame({"Feature": X.columns,"Score": X_kbest.scores_}).sort_values(by="Score", ascending=False)

print(headr("Feature scores"))
round(feature_scores,2)

> Una característica tiene demasiado peso y determina toda la clasificación

# Eliminamos columnas sin peso

# columns_to_drop = feature_scores[feature_scores['Score'] < 0.1]['Feature']
# final_data.drop(columns=columns_to_drop, inplace=True)
# tabl(final_data)

## Exportar 

final_data.to_csv('../data/preprocessed/preprocessed_data_v1.csv', index=False)
